Move EvalbFormatWriter as well
